refactor: move component spec in data
