+++ simaple/data/passive/patch.py
@@ -1,11 +1,9 @@
-import copy
 import math  # pylint: disable=W0611
 from typing import Union
 
 import pydantic
 
-from simaple.core import Stat
-from simaple.spec.patch import DFSTraversePatch, Patch
+from simaple.spec.patch import DFSTraversePatch
 
 
 class SkillLevelPatch(DFSTraversePatch):
@@ -43,25 +41,3 @@ class SkillLevelPatch(DFSTraversePatch):
             skill_level += self.combat_orders_level
 
         return skill_level
-
-
-class VSkillImprovementPatch(Patch):
-    improvements: dict[str, int] = pydantic.Field(default_factory=dict)
-
-    def apply(self, raw: dict) -> dict:
-        output = copy.deepcopy(raw)
-        try:
-            improvement_scale = output.pop("v_improvement")
-        except KeyError as e:
-            raise KeyError(
-                "VSkillImprovementPatch assigned but no `v_improvement`."
-            ) from e
-
-        previous_modifier = Stat.parse_obj(output.get("modifier", {}))
-        new_modifier = previous_modifier + Stat(
-            final_damage_multiplier=improvement_scale
-            * self.improvements.get(output["name"], 0)
-        )
-        output["modifier"] = new_modifier.short_dict()
-
-        return output
