+++ simaple/data/skill/resources/components/adventurer.yaml
