+++ simaple/data/skill/resources/components/archmagefb.yaml
