+++ simaple/data/skill/resources/components/common.yaml
