--- tests/job/resources/test_passive_skill.yaml
@@ -1,11 +1,0 @@
-kind: PassiveSkill 
-version: v1
-metadata:
-  group: test_group
-data:
-  passive_skill_enabled: true
-  combat_orders_enabled: true
-  default_skill_level: 10
-  stat:
-    attack_power: 3 * skill_level
-    INT: 5 * skill_level
